requirements.txt
asgiref==3.7.2
dj-database-url==2.0.0
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docutils==0.20.1
gunicorn==21.2.0
numpy==1.25.2
packaging==23.1
pandas==2.0.3
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
statistics==1.0.3.5
typing_extensions==4.7.1
tzdata==2023.3
whitenoise==6.5.0
